second round of refactorings - extracted interface
